Validate gtimelog file
Gtimelog silently re-order the entries on read.
This it very problematic as it does lead to:
- unaccounted time;
- misallocated time.

There is a solution though: detect and inform the user of this situation
and let him/her fix it.

Though this patch overrides a private method, I think it is better than
the status quo.